#35 - 코드 최적화 과정 진행
포스트 삭제시, 좋아요와 댓글에 대해 소프트 딜리트 쿼리를 직접 리포지토리 단에 작성해 주면서, 최적화 과정 진행.

+++ src/main/java/com/fc8/snsproject/domain/post/controller/PostController.java
@@ -89,11 +89,11 @@ public class PostController {
     }
 
     @GetMapping(path = "/{post-id}/likes")
-    public Response<Integer> likeCount(
+    public Response<Long> likeCount(
             @PathVariable(name = "post-id") Long postId,
             @AuthenticationPrincipal UserDetails userDetails
     ) {
-        Integer likeCount = postService.getLikeCount(postId);
+        Long likeCount = postService.getLikeCount(postId);
 
         return Response.success(likeCount);
     }
